Login and SignUp Text Fields renewed. Added and non-empty field verification.
I modified the style of the Create Account and Login EditText to a style that implements TextInputLayout TextInputEditText. This style makes the text field more customizable.
Now when trying to create an account, it checks if all the text fields are full, otherwise an error message is displayed below the text field.