Fixed bug with requesting permissions when selecting a photo
The first time the user granted write permissions to the application to access a photo, they were not being reflected, because how I was requesting permission was not adequate. I asked a colleague and he told me that I needed to have an implemented method called onRequestPermissionsResult, in which I could check if the permission was granted and if so, create the intent that goes to the photo gallery, and also I noticed on the internet that permissions are requested differently in an Activity and a Fragment.